Aggregatefunction any supports changelog
