+++ src/AggregateFunctions/Streaming/AggregateFunctionMax.cpp
@@ -1,4 +1,4 @@
-#include "HelpersMinMax.h"
+#include "HelpersMinMaxAny.h"
 
 #include <AggregateFunctions/AggregateFunctionFactory.h>
