+++ src/AggregateFunctions/registerAggregateFunctions.cpp
@@ -106,6 +106,7 @@ void registerAggregateFunctionAvgRetract(AggregateFunctionFactory & factory);
 void registerAggregateFunctionsMaxRetract(AggregateFunctionFactory & factory);
 void registerAggregateFunctionsMinRetract(AggregateFunctionFactory & factory);
 void registerAggregateFunctionMinMaxKRetract(AggregateFunctionFactory & factory);
+void registerAggregateFunctionsAnyRetract(AggregateFunctionFactory & factory);
 void registerAggregateFunctionGroupUniqArrayRetract(AggregateFunctionFactory & factory);
 
 #if USE_ARG_MIN_MAX_FUNCS
@@ -200,7 +201,7 @@ void registerAggregateFunctions()
         Streaming::registerAggregateFunctionsMinRetract(factory);
         Streaming::registerAggregateFunctionMinMaxKRetract(factory);
         Streaming::registerAggregateFunctionGroupUniqArrayRetract(factory);
-
+        Streaming::registerAggregateFunctionsAnyRetract(factory);
 
         #if USE_ARG_MIN_MAX_FUNCS
         Streaming::registerAggregateFunctionsArgMaxRetract(factory);
