+++ src/Interpreters/Streaming/SubstituteStreamingFunction.cpp
@@ -54,10 +54,12 @@ std::unordered_map<String, String> StreamingFunctionData::changelog_func_map = {
     {"min", "__min_retract"},
     {"arg_min", "__arg_min_retract"},
     {"arg_max", "__arg_max_retract"},
-    {"latest", ""},
-    {"earliest", ""},
-    {"first_value", ""},
-    {"last_value", ""},
+    {"any", "any_retract"},
+    {"any_last", "any_last_retract"},
+    {"latest", "latest_retract"},
+    {"earliest", "earliest_retract"},
+    {"first_value", "first_value_retract"},
+    {"last_value", "last_value_retract"},
     {"top_k", ""},
     {"min_k", "__min_k_retract"},
     {"max_k", "__max_k_retract"},
